Extract AttributeDefiner to hold the mass assignment initializer and attribute definition DSL
